% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{absseq_score}
\alias{absseq_score}
\title{absseq_score}
\usage{
absseq_score(y)
}
\arguments{
\item{y}{ABSSeq df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports an ABSSeq df for use by mitch
}
\keyword{ABSSeq}
\keyword{import}
\keyword{mitch}
