% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(x, genesets, minsetsize = 10, cores = detectCores() - 1,
  priority = NULL)
}
\arguments{
\item{x}{a single column numerical table with each column containing differential expression scores for a contrast.
Rownames must match genesets.}

\item{genesets}{lists of genes imported by the gmt_imprt function or similar.}

\item{minsetsize}{the minimum number of genes required in a set for it to be included in the statistical analysis.}

\item{cores}{the number of parallel threads for computation. Defaults to the number of cores present minus 1.}

\item{priority}{the prioritisation metric to selecting top gene sets. Valid options are 'significance' and 
'effect'.}
}
\value{
a mitch results object
}
\description{
This function performs the univariate analysis of variance for each set of genes. It wraps around the existing
aov function in R. This function is not meant to be used directly.
}
\examples{
#This function is not designed to be used directly
}
\keyword{MANOVA}
\keyword{mitch}
