% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{colname_substitute}
\alias{colname_substitute}
\title{colname_substitute}
\usage{
colname_substitute(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function substitutes colnames names in high dimensional data
}
\examples{
#This function is not designed to be used directly
}
\keyword{colnames}
\keyword{mitch}
