% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{deseq2_score}
\alias{deseq2_score}
\title{deseq2_score}
\usage{
deseq2_score(y)
}
\arguments{
\item{y}{DESeq2 df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a DESeq2 df for use by mitch
}
\keyword{deseq2}
\keyword{import}
\keyword{mitch}
