% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{detailed_sets}
\alias{detailed_sets}
\title{detailed_sets}
\usage{
detailed_sets(res, resrows = 50)
}
\arguments{
\item{res}{a mitch results object}

\item{resrows}{an integer representing the number of top genesets for which a detailed report is to be generated.
Default is 50.}
}
\value{
mitch res object with dataframes of the high priority gene sets attached.
}
\description{
This function fetches differential expression scores for members of sets of genes that were top ranked. This 
function is not meant to be used directly.
}
\examples{
#This function is not designed to be used directly
}
\keyword{detailed}
\keyword{mitch}
\keyword{sets}
