% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{edger_score}
\alias{edger_score}
\title{edger_score}
\usage{
edger_score(y)
}
\arguments{
\item{y}{edgeR df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports an edgeR df for use by mitch
}
\keyword{edgeR}
\keyword{import}
\keyword{mitch}
