% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{geneset_sector_table}
\alias{geneset_sector_table}
\title{geneset_sector_table}
\usage{
geneset_sector_table(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a table of gene sector counts
}
\examples{
#This function is not designed to be used directly
}
\keyword{mitch}
\keyword{sector}
\keyword{table}
