% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{genesets}
\alias{genesets}
\title{Reactome gene sets}
\format{A list of gene sets}
\source{
Reactome website: https://reactome.org/
}
\usage{
data(genesets)
}
\description{
Genesets from Reactome database suitable for enrichment analysis.
Acquired August 2019
}
\examples{
data(genesets)
}
\references{
Fabregat et al. (2017) BMC Bioinformatics volume 18, Article number: 142 
https://www.ncbi.nlm.nih.gov/pubmed/28249561
}
\keyword{datasets}
