% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{genesetsExample}
\alias{genesetsExample}
\title{Reactome gene sets}
\format{A list of gene sets}
\source{
Reactome website: https://reactome.org/
}
\usage{
data(genesetsExample)
}
\description{
Genesets from Reactome database suitable for enrichment analysis.
Acquired August 2019. The structure of this data is a named list of
vectors, containing human gene names (character strings). This is 
a sample of 200 gene sets from the approximately 2000 present in the
full dataset.
}
\examples{
data(genesetsExample)
}
\references{
Fabregat et al. (2017) BMC Bioinformatics volume 18, 
Article number: 142, https://www.ncbi.nlm.nih.gov/pubmed/28249561
}
\keyword{datasets}
