% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{ggpairs_points}
\alias{ggpairs_points}
\title{ggpairs_points}
\usage{
ggpairs_points(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a pairs plot of gene profiles
}
\examples{
#This function is not designed to be used directly
}
\keyword{enrichment}
\keyword{mitch}
\keyword{multidimensional}
\keyword{plots}
\keyword{scattter}
