% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{ggpairs_points_subset}
\alias{ggpairs_points_subset}
\title{ggpairs_points_subset}
\usage{
ggpairs_points_subset(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a pairs plot of gene profile
}
\examples{
#This function is not designed to be used directly
}
\keyword{enrichment}
\keyword{mitch}
\keyword{multidimensional}
\keyword{plots}
\keyword{scattter}
