% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{heatmapx}
\alias{heatmapx}
\title{heatmapx}
\usage{
heatmapx(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a pairs contour plot of gene profiles
}
\examples{
#This function is not designed to be used directly
}
\keyword{contour}
\keyword{density}
\keyword{enrichment}
\keyword{mitch}
\keyword{multidimensional}
\keyword{plots}
