% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{k36a}
\alias{k36a}
\title{H3K36ac profile}
\format{data frame}
\usage{
data(k36a)
}
\description{
Example edgeR result of differential ChIP-seq H3K36ac.
}
\examples{
data(k36a)
}
\keyword{datasets}
