% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{limma_score}
\alias{limma_score}
\title{limma_score}
\usage{
limma_score(y)
}
\arguments{
\item{y}{limma df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a limma df for use by mitch
}
\keyword{import}
\keyword{limma}
\keyword{mitch}
