% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mapGeneIds}
\alias{mapGeneIds}
\title{mapGeneIds}
\usage{
mapGeneIds(y, z)
}
\arguments{
\item{y}{DGE df}

\item{z}{a 2 column df of gene IDs}
}
\value{
a DGE df with mapped IDs
}
\description{
This function maps gene IDs present in the profiling data with those in the gene sets.
}
\keyword{ID}
\keyword{gene}
\keyword{import}
\keyword{mapping}
\keyword{mitch}
