% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mitch_metrics_calc1d}
\alias{mitch_metrics_calc1d}
\title{mitch_metrics_calc1d}
\usage{
mitch_metrics_calc1d(x, genesets, anova_result, minsetsize = 10)
}
\arguments{
\item{x}{a multicolumn numerical table with each column containing differential expression scores for a contrast.
Rownames must match genesets.}

\item{genesets}{lists of genes imported by the gmt_imprt function or similar.}

\item{anova_result}{a valid result of the ANOVA function}

\item{minsetsize}{the minimum number of genes required in a set for it to be included in the statistical analysis.}
}
\value{
a list of metrics
}
\description{
This function collects some metrics from the anova analysis. This function is not meant to be used directly.
}
\examples{
#This function is not designed to be used directly
}
\keyword{metrics}
\keyword{mitch}
