% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mitch_rank}
\alias{mitch_rank}
\title{mitch_rank}
\usage{
mitch_rank(x)
}
\arguments{
\item{x}{a multicolumn numerical table with each column containing differential expression scores for a contrast.}
}
\value{
a ranked table differential expression data.
}
\description{
This function performs zero-centred ranking of differential contrast omics data.
}
\examples{
#This function is not designed to be used directly
}
\keyword{mitch}
\keyword{rank}
\keyword{ranking}
