% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mitch_report}
\alias{mitch_report}
\title{mitch_report}
\usage{
mitch_report(res, outfile)
}
\arguments{
\item{res}{a mitch results object.}

\item{outfile}{the destination file for the html report. should contain
'html' suffix. Defaults to 
'Rplots.pdf'}
}
\value{
generates a HTML file containing enrichment plots.
}
\description{
This function generates an R markdown based html report containing tables and
several plots of mitch results 
The plots are in png format, so are not as high in resolution as compared to
the PDF generated by mitch_plots 
function. The number of detailed sets to generate is dictated by the resrows
set in the mitch_calc command.
}
\examples{
mitch_report(resExample,'outres2.html')
}
\keyword{html}
\keyword{knitr}
\keyword{markdown}
\keyword{mitch}
\keyword{report}
