% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{muscat_score}
\alias{muscat_score}
\title{muscat_score}
\usage{
muscat_score(y)
}
\arguments{
\item{y}{muscat df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a muscat df for use by mitch
}
\keyword{import}
\keyword{mitch}
\keyword{muscat}
