% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{myImportedData}
\alias{myImportedData}
\title{myImportedData: Example imported profiles}
\format{data frame}
\usage{
data(myImportedData)
}
\description{
Example of three edgeR profiles imported using mitch
}
\examples{
data(myImportedData)
}
\keyword{datasets}
