% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{myList}
\alias{myList}
\title{myList: A list of three edgeR results}
\format{
data frame
}
\usage{
data(myList)
}
\description{
Example edgeR results of differential RNA, H3K9ac and H3K36ac profiling.
The structure of this data is a list of three dataframes. Each data frame
is 1000 lines only.
}
\examples{
data(myList)
}
\keyword{datasets}
