% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{network_genes}
\alias{network_genes}
\title{network_genes}
\usage{
network_genes(eres, FDR = 0.05, n_sets = 20)
}
\arguments{
\item{eres}{a mitch results object.}

\item{FDR}{the significance threshold for inclusion. By default, it is 0.05}

\item{n_sets}{the number of sets to include.}
}
\value{
a list of up- and down-regulated gene sets
}
\description{
This is a companion function for the network diagrams to display the genes which contribute
to the enrichment and are shared between gene sets.
}
\examples{
data(resExample)
network_genes(resExample)
}
\keyword{mitch}
\keyword{network}
\keyword{visualisation}
