% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot1d_detailed}
\alias{plot1d_detailed}
\title{plot1d_detailed}
\usage{
plot1d_detailed(res, i)
}
\arguments{
\item{res}{a mitch results object}

\item{i}{the index of the geneset in the res$enrichment_result for plotting}
}
\description{
This function creates geneset histograms
}
\examples{
#This function is not designed to be used directly
}
\keyword{enrichment}
\keyword{mitch}
\keyword{plots}
\keyword{unidimensional}
