% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot1d_profile_dist}
\alias{plot1d_profile_dist}
\title{plot1d_profile_dist}
\usage{
plot1d_profile_dist(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates plots to describe input profile data for unidimensional enrichment analysis
}
\examples{
#This function is not designed to be used directly
}
\keyword{mitch}
\keyword{plots}
\keyword{unidimensional}
