% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot1d_volcano}
\alias{plot1d_volcano}
\title{plot1d_volcano}
\usage{
plot1d_volcano(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates geneset histograms
}
\examples{
#This function is not designed to be used directly
}
\keyword{mitch}
\keyword{plots}
\keyword{volcano}
