% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot2d_detailed_density}
\alias{plot2d_detailed_density}
\title{plot2d_detailed_density}
\usage{
plot2d_detailed_density(res, i)
}
\arguments{
\item{res}{a mitch results object}

\item{i}{the index of the geneset in the res$enrichment_result for plotting}
}
\description{
This function creates a 2D kernal density plot of genesetst
}
\examples{
#This function is not designed to be used directly
}
\keyword{bidimensional}
\keyword{enrichment}
\keyword{mitch}
\keyword{plots}
