% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot2d_profile_density}
\alias{plot2d_profile_density}
\title{plot2d_profile_density}
\usage{
plot2d_profile_density(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a kernal density plot for bidimensional enrichment analysis
}
\examples{
#This function is not designed to be used directly
}
\keyword{bidimensional}
\keyword{density}
\keyword{mitch}
\keyword{plots}
