% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot2d_profile_dist}
\alias{plot2d_profile_dist}
\title{plot2d_profile_dist}
\usage{
plot2d_profile_dist(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a scatterplot of input profile data for bidimensional enrichment analysis
}
\examples{
#This function is not designed to be used directly
}
\keyword{bidimensional}
\keyword{mitch}
\keyword{plots}
\keyword{scatterplot}
