% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot2d_set_quadrant_barchart}
\alias{plot2d_set_quadrant_barchart}
\title{plot2d_set_quadrant_barchart}
\usage{
plot2d_set_quadrant_barchart(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a barchart of genesets quadrants for bidimensional enrichment analysis
}
\examples{
#This function is not designed to be used directly
}
\keyword{barchart}
\keyword{barplot}
\keyword{bidimensional}
\keyword{mitch}
\keyword{plots}
