% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot2d_set_scatter}
\alias{plot2d_set_scatter}
\title{plot2d_set_scatter}
\usage{
plot2d_set_scatter(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a scatterplot of geneset enrichments
}
\examples{
#This function is not designed to be used directly
}
\keyword{bidimensional}
\keyword{mitch}
\keyword{plots}
\keyword{scatter}
