% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot3d_detailed_density}
\alias{plot3d_detailed_density}
\title{plot3d_detailed_density}
\usage{
plot3d_detailed_density(res, i)
}
\arguments{
\item{res}{a mitch results object}

\item{i}{the index of the geneset in the res$enrichment_result for plotting}
}
\description{
This function creates a pairs contour plot of gene profiles
}
\examples{
#This function is not designed to be used directly
}
\keyword{contour}
\keyword{density}
\keyword{enrichment}
\keyword{mitch}
\keyword{multidimensional}
\keyword{plots}
