% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot3d_detailed_points}
\alias{plot3d_detailed_points}
\title{plot3d_detailed_points}
\usage{
plot3d_detailed_points(res, i)
}
\arguments{
\item{res}{a mitch results object}

\item{i}{the index of the geneset in the res$enrichment_result for plotting}
}
\description{
This function creates a pairs plot of gene profiles of a subset of genes
}
\examples{
#This function is not designed to be used directly
}
\keyword{enrichment}
\keyword{mitch}
\keyword{multidimensional}
\keyword{pairs}
\keyword{plots}
\keyword{scatter}
