% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot3d_detailed_violin}
\alias{plot3d_detailed_violin}
\title{plot2d_detailed_violin}
\usage{
plot3d_detailed_violin(res, i)
}
\arguments{
\item{res}{a mitch results object}

\item{i}{the index of the geneset in the res$enrichment_result for plotting}
}
\description{
This function creates violin plots
}
\examples{
#This function is not designed to be used directly
}
\keyword{biidimensional}
\keyword{enrichment}
\keyword{mitch}
\keyword{plots}
\keyword{violin}
