% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot_effect_vs_significance}
\alias{plot_effect_vs_significance}
\title{plot_effect_vs_significance}
\usage{
plot_effect_vs_significance(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates a scatterplot of ES (effect) vs -log10(p.adjustMANOVA) (significance).
}
\examples{
#This function is not designed to be used directly
}
\keyword{effect}
\keyword{mitch}
\keyword{plots}
\keyword{significance}
