% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{plot_geneset_hist}
\alias{plot_geneset_hist}
\title{plot_geneset_hist}
\usage{
plot_geneset_hist(res)
}
\arguments{
\item{res}{a mitch results object}
}
\description{
This function creates geneset histograms
}
\examples{
#This function is not designed to be used directly
}
\keyword{histogram}
\keyword{mitch}
\keyword{plots}
