% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{resExample}
\alias{resExample}
\title{resExample: Example mitch result}
\format{list of mixed data types}
\usage{
data(resExample)
}
\description{
Example of mitch results. Enrichment of the Reactome gene sets in the RNA,
H3K9ac and H3K36ac datasets.
}
\examples{
data(resExample)
}
\keyword{datasets}
