% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{rna}
\alias{rna}
\title{RNA profile}
\format{data frame}
\usage{
data(rna)
}
\description{
Example edgeR result of differential RNA expression.
This is a dataframe which contains columns for log fold change, log counts
per million, p-value and FDR adjusted p-value. These columns consist of 
numerical values. The row names represent human gene names. This is a sample
of 1000 gene of an original dataset that contains measurements of ~15000 
genes.
}
\examples{
data(rna)
}
\keyword{datasets}
