% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{seurat_score}
\alias{seurat_score}
\title{seurat_score}
\usage{
seurat_score(y)
}
\arguments{
\item{y}{seurat df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a seurat df for use by mitch
}
\keyword{import}
\keyword{mitch}
\keyword{seurat}
