% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{sleuth_score}
\alias{sleuth_score}
\title{sleuth_score}
\usage{
sleuth_score(y)
}
\arguments{
\item{y}{sleuth df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a sleuth df for use by mitch
}
\keyword{import}
\keyword{mitch}
\keyword{sleuth}
