% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{topconfect_score}
\alias{topconfect_score}
\title{topconfect_score}
\usage{
topconfect_score(y)
}
\arguments{
\item{y}{topconfect df}
}
\value{
a single column df of ranked DE values
}
\description{
This function imports a topconfect df for use by mitch
}
\keyword{import}
\keyword{mitch}
\keyword{topconfect}
