% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{k36a}
\alias{k36a}
\title{H3K36ac profile}
\format{
data frame
}
\usage{
data(k36a)
}
\description{
Example edgeR result of differential ChIP-seq H3K36ac.
This is a dataframe which contains columns for log fold change, log counts
per million, p-value and FDR adjusted p-value. These columns consist of 
numerical values. The row names represent human gene names. This is a sample
of 1000 gene of an original dataset that contains measurements of ~30000 
genes.
}
\examples{
data(k36a)
}
\keyword{datasets}
