% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\name{mitch_plots}
\alias{mitch_plots}
\title{mitch_plots}
\usage{
mitch_plots(res, outfile = "Rplots.pdf")
}
\arguments{
\item{res}{a mitch results object.}

\item{outfile}{the destination file for the plots in PDF format. should
contain 'pdf' suffix. Defaults to 
'Rplots.pdf'}
}
\value{
generates a PDF file containing enrichment plots.
}
\description{
This function generates several plots of multivariate gene set enrichment in
high resolution PDF format.
The number of detailed sets to generate is dictated by the resrows set in the
mitch_calc command.
}
\examples{
data(resExample)
mitch_plots(resExample,outfile='outres.pdf')
}
\keyword{mitch}
\keyword{pdf}
\keyword{plot}
\keyword{plots}
