% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch.R
\docType{data}
\name{myImportedData}
\alias{myImportedData}
\title{myImportedData: Example imported profiles}
\format{
data frame
}
\usage{
data(myImportedData)
}
\description{
Example of three edgeR profiles imported using mitch.
The structure of this data is a dataframe where each column represents
one of the following profiling datasets after scoring: RNA, H3K9ac and
H3K36ac. Each row represents one gene and this dataset contains just 1000
rows to keep the example dataset small.
}
\examples{
data(myImportedData)
}
\keyword{datasets}
